+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add validation to require PURLs (OSIDB-4810)
 - Automatic affectedness and resolution determination based on product
   definitions (OSIDB-4938)
+- Added `bu_labels` label type (OSIDB-4982)
 
 ## [5.10.3] - 2026-05-06
 ### Fixed